requirements.txt
numpy==1.26.0
PySide6-Fluent-Widgets==1.7.0
loguru==0.7.2
requests==2.32.3
vispy==0.14.3
PyOpenGL==3.1.7